requirements.txt
opencv-python==4.5.3.56
torch==1.9.0
torchvision==0.10.0
onnxruntime==1.8.1
pyyaml==5.4.1
scipy==1.7.1
numpy==1.21.2
pandas==1.3.3
